requirements.txt
Flask==2.3.2
googletrans==4.0.0-rc1
PyPDF2==3.0.0
python-docx==0.8.11
reportlab==4.0.0
werkzeug==2.3.7
